Update substdo.c
Wrong parameter declaration, newer compiler are strict with parameters

+++ substdo.c
@@ -4,7 +4,7 @@
 #include "error.h"
 
 static int allwrite(op,fd,buf,len)
-register int (*op)();
+register ssize_t (*op)();
 register int fd;
 register char *buf;
 register int len;
